+++ src/os2datascanner/projects/admin/adminapp/models/scannerjobs/sbsysdb.py
@@ -21,7 +21,7 @@ class SBSYSDBScanner(Scanner):
 
     @classmethod
     def get_type(cls):
-        return "sbsys-db"
+        return "sbsys"
 
     @staticmethod
     def enabled():
@@ -116,5 +116,5 @@ class SBSYSDBScanner(Scanner):
                         spec_template.rule))
 
     class Meta:
-        verbose_name = _("SBSYS database scanner")
-        verbose_name_plural = _("SBSYS database scanners")
+        verbose_name = _("SBSYS scanner")
+        verbose_name_plural = _("SBSYS scanners")